@@ -2,20 +2,27 @@ plugins {
     id 'com.android.library'
     id 'kotlin-android'
 }
-apply plugin: 'kotlin-kapt'
 
-kapt {
-    correctErrorTypes = true
+if (project.hasProperty('useKsp')) {
+    apply plugin: 'com.google.devtools.ksp'
+} else {
+    apply plugin: 'kotlin-kapt'
+    kapt {
+        correctErrorTypes = true
+        arguments {
+            arg("multiPreviewType","com.airbnb.android.submodule.showkasesample.LocalePreview")
+            arg("multiPreviewType", "com.airbnb.android.submodule.showkasesample.FontPreview")
+        }
+    }
 }
 
 android {
-    compileSdkVersion 29
+    namespace "com.airbnb.android.submodule.showkasesample"
 
     defaultConfig {
         minSdkVersion 21
-        targetSdkVersion 29
-        versionCode 1
-        versionName "1.0"
+        compileSdk 34
+        targetSdkVersion 33
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
@@ -27,26 +34,31 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
     }
     kotlinOptions {
-        jvmTarget = '1.8'
+        jvmTarget = '17'
     }
     buildFeatures {
         compose true
     }
     composeOptions {
         kotlinCompilerVersion "${versions.kotlinCompilerVersion}"
-        kotlinCompilerExtensionVersion "${versions.compose}"
+        kotlinCompilerExtensionVersion "${versions.composeCompiler}"
 
+    }
+    // Added to avoid this error -
+    // Execution failed for task ':app:mergeDebugAndroidTestJavaResource'.
+    // > A failure occurred while executing com.android.build.gradle.internal.tasks.MergeJavaResWorkAction
+    // > 2 files found with path 'META-INF/AL2.0' from inputs:
+    packagingOptions {
+        exclude 'META-INF/AL2.0'
+        exclude 'META-INF/LGPL2.1'
     }
 }
 
 dependencies {
-    // Kotlin
-    implementation deps.kotlinStdLib
-    
     // Support Libraries
     implementation deps.support.appCompat
     implementation deps.support.ktx
@@ -54,10 +66,16 @@ dependencies {
     
     // Showkase
     implementation project(':showkase')
-    kapt project(':showkase-processor')
+    if (project.hasProperty('useKsp')) {
+        ksp project(':showkase-processor')
+    } else {
+        kapt project(':showkase-processor')
+    }
 
     // Compose
+    implementation deps.compose.activityCompose
     implementation deps.compose.composeRuntime
+    implementation deps.compose.constraintLayout
     implementation deps.compose.core
     implementation deps.compose.foundation
     implementation deps.compose.tooling
