@@ -1,34 +1,16 @@
-buildscript {
-    repositories {
-        mavenCentral()
-    }
-    // Maven artifact publish
-    dependencies {
-        classpath 'com.vanniktech:gradle-maven-publish-plugin:0.11.1'
-    }
-}
-
 plugins {
     id 'com.android.library'
     id 'kotlin-android'
-    id 'kotlin-android-extensions'
 }
-apply plugin: 'kotlin-kapt'
-apply plugin: "com.vanniktech.maven.publish"
-
-kapt {
-    correctErrorTypes = true
-    generateStubs = true
-}
+apply from: "$rootDir/publishing.gradle"
 
 android {
-    compileSdkVersion 29
+    namespace "com.airbnb.android.showkase"
 
     defaultConfig {
         minSdkVersion 21
-        targetSdkVersion 29
-        versionCode 1
-        versionName "1.0"
+        compileSdk 34
+        targetSdkVersion 33
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         consumerProguardFiles "consumer-rules.pro"
@@ -41,46 +23,47 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
     }
     kotlinOptions {
-        jvmTarget = '1.8'
+        jvmTarget = '17'
+        freeCompilerArgs += [
+                '-Xopt-in=kotlin.RequiresOptIn',
+        ]
     }
     buildFeatures {
         compose true
     }
     composeOptions {
         kotlinCompilerVersion "${versions.kotlinCompilerVersion}"
-        kotlinCompilerExtensionVersion "${versions.compose}"
+        kotlinCompilerExtensionVersion "${versions.composeCompiler}"
+    }
+    // Added to avoid this error -
+    // Execution failed for task ':app:mergeDebugAndroidTestJavaResource'.
+    // > A failure occurred while executing com.android.build.gradle.internal.tasks.MergeJavaResWorkAction
+    // > 2 files found with path 'META-INF/AL2.0' from inputs:
+    packagingOptions {
+        exclude 'META-INF/AL2.0'
+        exclude 'META-INF/LGPL2.1'
     }
 }
 
 dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    // Kotlin
-    implementation deps.kotlinStdLib
-
     // Showkase
     api project(':showkase-annotation')
 
     // Support Libraries
     implementation deps.support.appCompat
     implementation deps.support.ktx
-    implementation deps.support.lifecycleExtensions
-    
+
     // Compose
+    implementation deps.compose.activityCompose
     implementation deps.compose.composeRuntime
+    implementation deps.compose.composeNavigation
     implementation deps.compose.core
     implementation deps.compose.foundation
     implementation deps.compose.tooling
     implementation deps.compose.layout
     implementation deps.compose.material
-    implementation deps.compose.savedInstanceState
-    implementation deps.compose.uiLiveData
-    androidTestImplementation deps.compose.uiTest
-    
-    // Unit Testing
-    testImplementation deps.test.junit
-    androidTestImplementation deps.test.junitImplementation
 }
