@@ -5,19 +5,19 @@ import android.graphics.drawable.Drawable
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.onCommit
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.WithConstraints
-import androidx.compose.ui.graphics.ImageAsset
-import androidx.compose.ui.graphics.asImageAsset
+import androidx.compose.ui.graphics.ImageBitmap
+import androidx.compose.ui.graphics.asImageBitmap
 import androidx.compose.ui.graphics.drawscope.drawIntoCanvas
 import androidx.compose.ui.graphics.nativeCanvas
 import com.squareup.picasso.Picasso
@@ -26,18 +26,18 @@ import com.squareup.picasso.Target
 @Composable
 fun NetworkImage(
     imageUrl: String,
-    modifier: Modifier = Modifier.fillMaxWidth()
+    modifier: Modifier = Modifier
 ) {
     Column(
-        modifier = modifier,
+        modifier = modifier.fillMaxWidth(),
         verticalArrangement = Arrangement.Center,
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
-        WithConstraints {
+        BoxWithConstraints {
             val constrainScope = this
-            var image by remember { mutableStateOf<ImageAsset?>(null) }
+            var image by remember { mutableStateOf<ImageBitmap?>(null) }
             var drawable by remember { mutableStateOf<Drawable?>(null) }
-            onCommit(imageUrl) {
+            DisposableEffect(imageUrl) {
                 val picasso = Picasso.get()
                 val target = object : Target {
                     override fun onPrepareLoad(placeHolderDrawable: Drawable?) {
@@ -49,7 +49,7 @@ fun NetworkImage(
                     }
 
                     override fun onBitmapLoaded(bitmap: Bitmap?, from: Picasso.LoadedFrom?) {
-                        image = bitmap?.asImageAsset()
+                        image = bitmap?.asImageBitmap()
                     }
                 }
 
@@ -69,7 +69,7 @@ fun NetworkImage(
             val theImage = image
             val theDrawable = drawable
             if (theImage != null) {
-                Image(asset = theImage)
+                Image(bitmap = theImage, contentDescription = null)
             } else if (theDrawable != null) {
                 Canvas(modifier = modifier) {
                     drawIntoCanvas { canvas ->
