@@ -1,14 +1,16 @@
 package com.airbnb.android.showkasesample
 
 import android.os.Bundle
+import androidx.activity.compose.setContent
 import androidx.appcompat.app.AppCompatActivity
-import androidx.compose.ui.platform.setContent
+import com.airbnb.android.showkase.models.Showkase
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            startActivity(createShowkaseBrowserIntent(this))
+            startActivity(Showkase.getBrowserIntent(this))
+            finish()
         }
     }
 }
