@@ -4,4 +4,4 @@ import com.airbnb.android.showkase.annotation.ShowkaseRoot
 import com.airbnb.android.showkase.annotation.ShowkaseRootModule
 
 @ShowkaseRoot
-class RootModule: ShowkaseRootModule
+class RootModule : ShowkaseRootModule
