@@ -7,7 +7,7 @@ import androidx.compose.material.lightColors
 import androidx.compose.runtime.Composable
 
 @Composable
-fun ShowkaseTheme(children: @Composable()() -> Unit) {
+fun ShowkaseTheme(children: @Composable () -> Unit) {
     val light = lightColors()
     val dark = darkColors()
     val colors = if (isSystemInDarkTheme()) { dark } else { light }
