@@ -1,252 +1,224 @@
 package com.airbnb.android.showkase.processor.writer
 
-import com.airbnb.android.showkase.processor.exceptions.ShowkaseProcessorException
-import com.airbnb.android.showkase.processor.models.ShowkaseMetadata
+import androidx.room.compiler.processing.XElement
+import androidx.room.compiler.processing.XFiler
+import androidx.room.compiler.processing.XProcessingEnv
+import androidx.room.compiler.processing.addOriginatingElement
+import androidx.room.compiler.processing.get
+import androidx.room.compiler.processing.isTypeElement
+import androidx.room.compiler.processing.writeTo
+import com.airbnb.android.showkase.annotation.ShowkaseMultiPreviewCodegenMetadata
+import com.airbnb.android.showkase.annotation.ShowkaseRootCodegen
+import com.airbnb.android.showkase.processor.ShowkaseGeneratedMetadata
+import com.airbnb.android.showkase.processor.ShowkaseProcessor
+import com.squareup.kotlinpoet.AnnotationSpec
 import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.CodeBlock
+import com.squareup.kotlinpoet.FileSpec
+import com.squareup.kotlinpoet.FunSpec
+import com.squareup.kotlinpoet.LIST
 import com.squareup.kotlinpoet.MemberName
-import com.squareup.kotlinpoet.PropertySpec
-import javax.annotation.processing.ProcessingEnvironment
-import javax.lang.model.type.TypeMirror
+import com.squareup.kotlinpoet.ParameterizedTypeName.Companion.parameterizedBy
+import com.squareup.kotlinpoet.TypeSpec
+import java.util.Locale
 
-internal class ShowkaseBrowserWriter(private val processingEnv: ProcessingEnvironment) {
-    @Suppress("LongMethod")
+internal class ShowkaseBrowserWriter(private val environment: XProcessingEnv) {
+    @Suppress("LongMethod", "LongParameterList")
     internal fun generateShowkaseBrowserFile(
-        showkaseComponentMetadata: Set<ShowkaseMetadata>,
-        showkaseColorMetadata: Set<ShowkaseMetadata>,
-        showkaseTypographyMetadata: Set<ShowkaseMetadata>,
+        allShowkaseBrowserProperties: ShowkaseBrowserProperties,
         rootModulePackageName: String,
         rootModuleClassName: String
     ) {
-        val showkaseComponentsListClassName = "$rootModuleClassName$CODEGEN_AUTOGEN_CLASS_NAME"
+        val showkaseComponentsListClassName = "${rootModuleClassName}$CODEGEN_AUTOGEN_CLASS_NAME"
         val fileBuilder = getFileBuilder(rootModulePackageName, showkaseComponentsListClassName)
+        val componentCodeBlock = initializeComponentCodeBlock(
+            allShowkaseBrowserProperties.componentsWithoutPreviewParameters,
+            allShowkaseBrowserProperties.componentsWithPreviewParameters
+        )
+        val colorCodeBlock = initializeColorCodeBlock(allShowkaseBrowserProperties.colors)
+        val typographyCodeBlock =
+            initializeTypographyCodeBlock(allShowkaseBrowserProperties.typography)
 
-        val componentListProperty = initializeComponentProperty(showkaseComponentMetadata)
-        val colorListProperty = initializeColorProperty(showkaseColorMetadata)
-        val typographyProperty = initializeTypographyProperty(showkaseTypographyMetadata)
+        val showkaseRootCodegenAnnotation = initializeShowkaseRootCodegenAnnotation(
+            allShowkaseBrowserProperties.componentsWithoutPreviewParameters.size,
+            allShowkaseBrowserProperties.componentsWithPreviewParameters.size,
+            allShowkaseBrowserProperties.colors.size,
+            allShowkaseBrowserProperties.typography.size
+        )
 
         writeFile(
-            processingEnv,
+            environment,
             fileBuilder,
             SHOWKASE_PROVIDER_CLASS_NAME,
             showkaseComponentsListClassName,
-            componentListProperty.build(),
-            colorListProperty.build(),
-            typographyProperty.build(),
-            showkaseComponentMetadata + showkaseColorMetadata + showkaseTypographyMetadata,
+            allShowkaseBrowserProperties,
             getShowkaseProviderInterfaceFunction(
-                COMPONENT_INTERFACE_METHOD_NAME,
-                COMPONENT_PROPERTY_NAME
+                methodName = COMPONENT_INTERFACE_METHOD_NAME,
+                returnType = LIST.parameterizedBy(SHOWKASE_BROWSER_COMPONENT_CLASS_NAME),
+                codeBlock = componentCodeBlock
             ),
             getShowkaseProviderInterfaceFunction(
-                COLOR_INTERFACE_METHOD_NAME,
-                COLOR_PROPERTY_NAME
+                methodName = COLOR_INTERFACE_METHOD_NAME,
+                returnType = LIST.parameterizedBy(SHOWKASE_BROWSER_COLOR_CLASS_NAME),
+                codeBlock = colorCodeBlock
             ),
             getShowkaseProviderInterfaceFunction(
-                TYPOGRAPHY_INTERFACE_METHOD_NAME,
-                TYPOGRAPHY_PROPERTY_NAME
-            )
+                methodName = TYPOGRAPHY_INTERFACE_METHOD_NAME,
+                returnType = LIST.parameterizedBy(SHOWKASE_BROWSER_TYPOGRAPHY_CLASS_NAME),
+                codeBlock = typographyCodeBlock
+            ),
+            showkaseRootCodegenAnnotation
         )
     }
 
-    private fun initializeComponentProperty(
-        showkaseMetadataSet: Set<ShowkaseMetadata>
-    ): PropertySpec.Builder {
-        val componentListProperty =
-            getPropertyList(SHOWKASE_BROWSER_COMPONENT_CLASS_NAME, COMPONENT_PROPERTY_NAME)
-
-        val componentListInitializerCodeBlock =
+    private fun initializeComponentCodeBlock(
+        withoutParameterPropertyNames: List<ShowkaseGeneratedMetadata>,
+        withParameterPropertyNames: List<ShowkaseGeneratedMetadata>,
+    ): CodeBlock {
+        val componentListInitializerCodeBlock = if (withParameterPropertyNames.isNotEmpty()) {
             SHOWKASE_BROWSER_COMPONENT_CLASS_NAME.mutableListInitializerCodeBlock()
-        val (showkaseMetadataWithParameterList, showkaseMetadataWithoutParameterList) =
-            showkaseMetadataSet.filterIsInstance<ShowkaseMetadata.Component>().partition {
-                it.previewParameter != null
-            }
+        } else {
+            SHOWKASE_BROWSER_COMPONENT_CLASS_NAME.listInitializerCodeBlock()
+        }
 
         componentListInitializerCodeBlock.apply {
-            showkaseMetadataWithoutParameterList.forEachIndexed { index, withoutParameterMetadata ->
+            addLineBreak()
+            withoutParameterPropertyNames.forEachIndexed { index, metadata ->
+                add("%M,", MemberName(metadata.propertyPackage, metadata.propertyName))
                 addLineBreak()
-                addShowkaseBrowserComponent(withoutParameterMetadata)
-                closeOrContinueListCodeBlock(index, showkaseMetadataWithoutParameterList.size - 1)
             }
-            unindent()
+            doubleUnindent()
             add(")")
 
-            if (showkaseMetadataWithParameterList.isNotEmpty()) {
+            if (withParameterPropertyNames.isNotEmpty()) {
                 add(".apply {")
-                doubleIndent()
-                showkaseMetadataWithParameterList.forEachIndexed { _, withParameterMetadata ->
-                    addLineBreak()
-                    add(
-                        "%T().values.iterator().forEach { previewParam -> ",
-                        withParameterMetadata.previewParameter
-                    )
-                    doubleIndent()
-                    addLineBreak()
-                    add("add(")
-                    addLineBreak()
-                    doubleIndent()
-                    addShowkaseBrowserComponent(withParameterMetadata)
-                    closeRoundBracket()
-                    doubleUnindent()
-                    closeRoundBracket()
-                    doubleUnindent()
-                    closeCurlyBraces()
+                addLineBreak()
+                withDoubleIndent {
+                    withParameterPropertyNames.forEachIndexed { index, metadata ->
+                        add(
+                            "addAll(%M)",
+                            MemberName(metadata.propertyPackage, metadata.propertyName)
+                        )
+                        if (index != withParameterPropertyNames.lastIndex) {
+                            addLineBreak()
+                        }
+                    }
                 }
-                doubleUnindent()
                 closeCurlyBraces()
             }
         }
-        
-        componentListProperty.initializer(componentListInitializerCodeBlock.build())
-        return componentListProperty
+
+        return componentListInitializerCodeBlock.build()
     }
 
-    private fun initializeColorProperty(
-        showkaseMetadataSet: Set<ShowkaseMetadata>
-    ): PropertySpec.Builder {
-        val colorListProperty =
-            getPropertyList(
-                SHOWKASE_BROWSER_COLOR_CLASS_NAME,
-                COLOR_PROPERTY_NAME
-            )
-
+    private fun initializeColorCodeBlock(
+        colorsParameterPropertyNames: List<ShowkaseGeneratedMetadata>,
+    ): CodeBlock {
         val colorListInitializerCodeBlock =
             SHOWKASE_BROWSER_COLOR_CLASS_NAME.listInitializerCodeBlock()
 
-        showkaseMetadataSet.forEachIndexed { index, showkaseMetadata ->
-            colorListInitializerCodeBlock.apply {
+        return colorListInitializerCodeBlock.apply {
+            addLineBreak()
+            colorsParameterPropertyNames.forEachIndexed { index, metadata ->
+                add("%M,", MemberName(metadata.propertyPackage, metadata.propertyName))
                 addLineBreak()
-                add(
-                    "%T(\n",
-                    SHOWKASE_BROWSER_COLOR_CLASS_NAME
-                )
-                doubleIndent()
-                add(
-                    "colorGroup = %S,\ncolorName = %S,\ncolorKDoc = %S,",
-                    showkaseMetadata.showkaseGroup,
-                    showkaseMetadata.showkaseName,
-                    showkaseMetadata.showkaseKDoc
-                )
-                add(
-                    showkaseBrowserPropertyValue(
-                        showkaseMetadata.packageName,
-                        showkaseMetadata.enclosingClass,
-                        "color",
-                        showkaseMetadata.elementName,
-                        showkaseMetadata.insideWrapperClass,
-                        showkaseMetadata.insideObject
-                    )
-                )
-                doubleUnindent()
-                closeOrContinueListCodeBlock(index, showkaseMetadataSet.size - 1)
             }
-        }
-        colorListInitializerCodeBlock.apply {
-            unindent()
+            doubleUnindent()
             add(")")
-        }
-        colorListProperty.initializer(colorListInitializerCodeBlock.build())
-        return colorListProperty
+        }.build()
     }
 
-    private fun initializeTypographyProperty(
-        showkaseMetadataSet: Set<ShowkaseMetadata>
-    ): PropertySpec.Builder {
-        val typographyListProperty =
-            getPropertyList(
-                SHOWKASE_BROWSER_TYPOGRAPHY_CLASS_NAME,
-                TYPOGRAPHY_PROPERTY_NAME
-            )
-
+    private fun initializeTypographyCodeBlock(
+        typographyParameterPropertyNames: List<ShowkaseGeneratedMetadata>,
+    ): CodeBlock {
         val typographyListInitializerCodeBlock =
             SHOWKASE_BROWSER_TYPOGRAPHY_CLASS_NAME.listInitializerCodeBlock()
 
-        showkaseMetadataSet.forEachIndexed { index, showkaseMetadata ->
-            typographyListInitializerCodeBlock.apply {
+        return typographyListInitializerCodeBlock.apply {
+            addLineBreak()
+            typographyParameterPropertyNames.forEachIndexed { index, metadata ->
+                add("%M,", MemberName(metadata.propertyPackage, metadata.propertyName))
                 addLineBreak()
-                add(
-                    "%T(\n",
-                    SHOWKASE_BROWSER_TYPOGRAPHY_CLASS_NAME
-                )
-                doubleIndent()
-                add(
-                    "typographyGroup = %S,\ntypographyName = %S,\ntypographyKDoc = %S,",
-                    showkaseMetadata.showkaseGroup,
-                    showkaseMetadata.showkaseName,
-                    showkaseMetadata.showkaseKDoc
-                )
-                add(
-                    showkaseBrowserPropertyValue(
-                        showkaseMetadata.packageName,
-                        showkaseMetadata.enclosingClass,
-                        "textStyle",
-                        showkaseMetadata.elementName,
-                        showkaseMetadata.insideWrapperClass,
-                        showkaseMetadata.insideObject
-                    )
-                )
-                doubleUnindent()
-                closeOrContinueListCodeBlock(index, showkaseMetadataSet.size - 1)
             }
-        }
-        typographyListInitializerCodeBlock.apply {
-            unindent()
+            doubleUnindent()
             add(")")
-        }
-        return typographyListProperty.initializer(typographyListInitializerCodeBlock.build())
+        }.build()
     }
 
-    @Suppress("LongParameterList")
-    private fun showkaseBrowserPropertyValue(
-        functionPackageName: String,
-        enclosingClass: TypeMirror? = null,
-        fieldPropertyName: String,
-        fieldName: String,
-        insideWrapperClass: Boolean,
-        insideObject: Boolean
-    ) = when {
-        // When enclosingClass is null, it denotes that the method was a top-level method 
-        // declaration.
-        enclosingClass == null -> {
-            val composeMember = MemberName(functionPackageName, fieldName)
-            CodeBlock.Builder()
-                .add("\n$fieldPropertyName = %M", composeMember)
-                .build()
+    private fun initializeShowkaseRootCodegenAnnotation(
+        numComponentsWithoutPreviewParameter: Int,
+        numComponentsWithPreviewParameter: Int,
+        colorsSize: Int,
+        typographySize: Int,
+    ) = AnnotationSpec.builder(ShowkaseRootCodegen::class)
+        .addMember(
+            "numComposablesWithoutPreviewParameter = %L",
+            numComponentsWithoutPreviewParameter
+        )
+        .addMember("numComposablesWithPreviewParameter = %L", numComponentsWithPreviewParameter)
+        .addMember("numColors = %L", colorsSize)
+        .addMember("numTypography = %L", typographySize)
+        .build()
+
+    // This is to aggregate metadata for the custom annotation annotated with Preview
+    internal fun writeCustomAnnotationElementToMetadata(element: XElement) {
+        if (!element.isTypeElement()) return
+        if (element.isAnnotationClass() && element.qualifiedName == ShowkaseProcessor.PREVIEW_CLASS_NAME) return
+
+        val moduleName = "Showkase_${element.qualifiedName.replace(".", "_")}"
+        val generatedClassName =
+            "ShowkaseMetadata_${moduleName.lowercase(Locale.getDefault())}"
+
+        val previewAnnotations =
+            element.getAllAnnotations().filter { it.name == ShowkaseProcessor.PREVIEW_SIMPLE_NAME }
+
+        val fileBuilder = FileSpec.builder(
+            ShowkaseProcessor.CODEGEN_PACKAGE_NAME,
+            generatedClassName
+        )
+
+        val functions = previewAnnotations.mapIndexed { index, xAnnotation ->
+            FunSpec.builder("${xAnnotation.name}_$index")
+                .addAnnotation(
+                    AnnotationSpec
+                        .builder(ShowkaseMultiPreviewCodegenMetadata::class)
+                        .addMember("previewName = %S", xAnnotation.get("name"))
+                        .addMember("previewGroup = %S", xAnnotation.get("group"))
+                        .addMember("supportTypeQualifiedName = %S", element.qualifiedName)
+                        .addMember("showkaseWidth = %L", xAnnotation.getAsInt("widthDp"))
+                        .addMember("showkaseHeight = %L", xAnnotation.getAsInt("heightDp"))
+                        .addMember("packageName = %S", element.packageName)
+                        .build()
+                ).build()
         }
-        // It was declared inside a class.
-        insideWrapperClass -> {
-            CodeBlock.Builder()
-                .add("\n$fieldPropertyName = %T().${fieldName}", enclosingClass)
-                .build()
-        }
-        // It was declared inside an object or a companion object.
-        insideObject -> {
-            CodeBlock.Builder()
-                .add("\n$fieldPropertyName = %T.${fieldName}", enclosingClass)
-                .build()
-        }
-        else -> throw ShowkaseProcessorException("Your field:${fieldName} is declared in a way that " +
-                "is not supported by Showkase")
+
+        fileBuilder.addType(
+            with(TypeSpec.classBuilder(generatedClassName).addFunctions(functions)) {
+                addOriginatingElement(element)
+                build()
+            }
+        ).addFileComment("This is an auto-generated file. Please do not edit/modify this file.")
+        fileBuilder.build().writeTo(environment.filer, mode = XFiler.Mode.Aggregating)
     }
 
     companion object {
-        private const val CODEGEN_AUTOGEN_CLASS_NAME = "Codegen"
+        internal const val CODEGEN_AUTOGEN_CLASS_NAME = "Codegen"
         private const val COMPONENT_INTERFACE_METHOD_NAME = "getShowkaseComponents"
         private const val COLOR_INTERFACE_METHOD_NAME = "getShowkaseColors"
         private const val TYPOGRAPHY_INTERFACE_METHOD_NAME = "getShowkaseTypography"
         internal const val SHOWKASE_MODELS_PACKAGE_NAME = "com.airbnb.android.showkase.models"
-        private const val COMPONENT_PROPERTY_NAME = "componentList"
-        private const val COLOR_PROPERTY_NAME = "colorList"
-        private const val TYPOGRAPHY_PROPERTY_NAME = "typographyList"
+        internal const val COMPONENT_PROPERTY_NAME = "componentList"
+        internal const val COLOR_PROPERTY_NAME = "colorList"
+        internal const val TYPOGRAPHY_PROPERTY_NAME = "typographyList"
 
-        val COMPOSE_CLASS_NAME = ClassName("androidx.compose.runtime", "Composable")
-        val SHOWKASE_BROWSER_COMPONENT_CLASS_NAME =
+        internal val COMPOSE_CLASS_NAME = ClassName("androidx.compose.runtime", "Composable")
+        internal val SHOWKASE_BROWSER_COMPONENT_CLASS_NAME =
             ClassName(SHOWKASE_MODELS_PACKAGE_NAME, "ShowkaseBrowserComponent")
-        val SHOWKASE_BROWSER_COLOR_CLASS_NAME =
+        internal val SHOWKASE_BROWSER_COLOR_CLASS_NAME =
             ClassName(SHOWKASE_MODELS_PACKAGE_NAME, "ShowkaseBrowserColor")
-        val SHOWKASE_BROWSER_TYPOGRAPHY_CLASS_NAME =
+        internal val SHOWKASE_BROWSER_TYPOGRAPHY_CLASS_NAME =
             ClassName(SHOWKASE_MODELS_PACKAGE_NAME, "ShowkaseBrowserTypography")
-        val SHOWKASE_PROVIDER_CLASS_NAME =
+        internal val SHOWKASE_PROVIDER_CLASS_NAME =
             ClassName(SHOWKASE_MODELS_PACKAGE_NAME, "ShowkaseProvider")
     }
 }
