@@ -1,5 +1,9 @@
 package com.airbnb.android.showkase.processor.writer
 
+import androidx.room.compiler.processing.XFiler
+import androidx.room.compiler.processing.XProcessingEnv
+import androidx.room.compiler.processing.addOriginatingElement
+import androidx.room.compiler.processing.writeTo
 import com.airbnb.android.showkase.annotation.ShowkaseCodegenMetadata
 import com.airbnb.android.showkase.processor.ShowkaseProcessor.Companion.CODEGEN_PACKAGE_NAME
 import com.airbnb.android.showkase.processor.models.ShowkaseMetadata
@@ -9,38 +13,45 @@ import com.squareup.kotlinpoet.FileSpec
 import com.squareup.kotlinpoet.FunSpec
 import com.squareup.kotlinpoet.TypeSpec
 import java.util.Locale
-import javax.annotation.processing.ProcessingEnvironment
-import javax.lang.model.util.Types
 
-internal class ShowkaseCodegenMetadataWriter(private val processingEnv: ProcessingEnvironment) {
+internal class ShowkaseCodegenMetadataWriter(private val environment: XProcessingEnv) {
 
     internal fun generateShowkaseCodegenFunctions(
         showkaseMetadataSet: Set<ShowkaseMetadata>,
-        typeUtil: Types
     ) {
-        if (showkaseMetadataSet.isEmpty()) return
-        val moduleName = showkaseMetadataSet.first().packageSimpleName
-        val generatedClassName = "ShowkaseMetadata${moduleName.capitalize(Locale.getDefault())}"
+        val moduleName = showkaseMetadataSet.first().packageName.replace(".", "_")
+        val generatedClassName = "ShowkaseMetadata_${moduleName.lowercase(Locale.getDefault())}"
         val fileBuilder = FileSpec.builder(
             CODEGEN_PACKAGE_NAME,
             generatedClassName
         )
-            .addComment("This is an auto-generated file. Please do not edit/modify this file.")
+            .addFileComment("This is an auto-generated file. Please do not edit/modify this file.")
 
         val autogenClass = TypeSpec.classBuilder(generatedClassName)
 
         showkaseMetadataSet.forEach { showkaseMetadata ->
-            val methodName = when {
-                showkaseMetadata.enclosingClass == null -> showkaseMetadata.elementName
-                else -> {
-                    val enclosingClassName =
-                        typeUtil.asElement(showkaseMetadata.enclosingClass).simpleName
-                    "${enclosingClassName}_${showkaseMetadata.elementName}"
-                }
+
+            val name = if (
+                showkaseMetadata is ShowkaseMetadata.Component &&
+                showkaseMetadata.componentIndex != null &&
+                showkaseMetadata.componentIndex > 0
+            ) {
+                "${showkaseMetadata.fqPrefix}_${showkaseMetadata.showkaseGroup}" +
+                    "_${showkaseMetadata.showkaseName}_${showkaseMetadata.componentIndex}"
+            } else {
+                "${showkaseMetadata.fqPrefix}_${showkaseMetadata.showkaseGroup}" +
+                    "_${showkaseMetadata.showkaseName}"
             }
+            val methodName = if (showkaseMetadata is ShowkaseMetadata.Component &&
+                showkaseMetadata.showkaseStyleName != null
+            ) {
+                "${name}_${showkaseMetadata.showkaseStyleName}"
+            } else {
+                name
+            }.filter { it.isLetterOrDigit() }
 
             val annotation = createShowkaseCodegenMetadata(showkaseMetadata)
-            showkaseMetadata.enclosingClass?.let {
+            showkaseMetadata.enclosingClassName?.let {
                 annotation.addMember("enclosingClass = [%T::class]", it)
             }
             addMetadataTypeSpecificProperties(showkaseMetadata, annotation)
@@ -58,10 +69,10 @@ internal class ShowkaseCodegenMetadataWriter(private val processingEnv: Processi
             }
         )
 
-        fileBuilder.build().writeTo(processingEnv.filer)
+        fileBuilder.build().writeTo(environment.filer, mode = XFiler.Mode.Aggregating)
     }
 
-    private fun createShowkaseCodegenMetadata(showkaseMetadata: ShowkaseMetadata) =
+    private fun createShowkaseCodegenMetadata(showkaseMetadata: ShowkaseMetadata): AnnotationSpec.Builder =
         AnnotationSpec.builder(ShowkaseCodegenMetadata::class)
             .addMember("showkaseName = %S", showkaseMetadata.showkaseName)
             .addMember("showkaseGroup = %S", showkaseMetadata.showkaseGroup)
@@ -71,6 +82,7 @@ internal class ShowkaseCodegenMetadataWriter(private val processingEnv: Processi
             .addMember("insideObject = ${showkaseMetadata.insideObject}")
             .addMember("insideWrapperClass = ${showkaseMetadata.insideWrapperClass}")
             .addMember("showkaseKDoc = %S", showkaseMetadata.showkaseKDoc)
+            .addMember("generatedPropertyName = %S", generatePropertyNameFromMetadata(showkaseMetadata))
 
     private fun addMetadataTypeSpecificProperties(
         showkaseMetadata: ShowkaseMetadata,
@@ -79,15 +91,24 @@ internal class ShowkaseCodegenMetadataWriter(private val processingEnv: Processi
         is ShowkaseMetadata.Component -> {
             annotation.apply {
                 addMember("showkaseMetadataType = %S", ShowkaseMetadataType.COMPONENT.name)
+                addMember("isDefaultStyle = ${showkaseMetadata.isDefaultStyle}")
                 showkaseMetadata.showkaseWidthDp?.let {
                     addMember("showkaseWidthDp = %L", it)
                 }
                 showkaseMetadata.showkaseHeightDp?.let {
                     addMember("showkaseHeightDp = %L", it)
                 }
-                showkaseMetadata.previewParameter?.let {
+                showkaseMetadata.previewParameterProviderType?.let {
                     addMember("previewParameterClass = [%T::class]", it)
                 }
+                showkaseMetadata.previewParameterName?.let {
+                    addMember("previewParameterName = %S", it)
+                }
+                showkaseMetadata.showkaseStyleName?.let {
+                    addMember("showkaseStyleName = %S", showkaseMetadata.showkaseStyleName)
+                }
+                addStringArrayMember(ShowkaseCodegenMetadata::tags.name, showkaseMetadata.tags)
+                addStringArrayMember(ShowkaseCodegenMetadata::extraMetadata.name, showkaseMetadata.extraMetadata)
             }
         }
         is ShowkaseMetadata.Color -> {
@@ -95,6 +116,15 @@ internal class ShowkaseCodegenMetadataWriter(private val processingEnv: Processi
         }
         is ShowkaseMetadata.Typography -> {
             annotation.addMember("showkaseMetadataType = %S", ShowkaseMetadataType.TYPOGRAPHY.name)
+        }
+    }
+
+    private fun AnnotationSpec.Builder.addStringArrayMember(name: String, values: List<String>) {
+        val valueAsArray = values.joinToString(", ", prefix = "[", postfix = "]") { value ->
+            "\"$value\""
+        }
+        values.takeIf { it.isNotEmpty() }?.let {
+            addMember("%L = %L", name, valueAsArray)
         }
     }
 }
