@@ -3,6 +3,6 @@ package com.airbnb.android.showkase.exceptions
 import java.lang.Exception
 
 /**
- * Used to throw an exception for Showkase specific errors. 
+ * Used to throw an exception for Showkase specific errors.
  */
-internal class ShowkaseException(message: String): Exception(message)
+internal class ShowkaseException(message: String) : Exception(message)
