@@ -1,6 +1,6 @@
 package com.airbnb.android.showkase.models
 
-internal enum class ShowkaseCategory {
+enum class ShowkaseCategory {
     COMPONENTS,
     COLORS,
     TYPOGRAPHY
